+++ partitioned-pipe-multiscale/fluid1d-right-nutils/clean.sh
@@ -4,5 +4,7 @@ set -e -u
 . ../../tools/cleaning-tools.sh
 
 rm -f ./results/Fluid1D_*
+rm -f ./probes.txt
+rm -f ./final_fields.txt
 clean_precice_logs .
 clean_case_logs .
