+++ partitioned-pipe-multiscale/fluid1d-right-nutils/requirements.txt
@@ -1,5 +1,5 @@
 setuptools # required by nutils
-nutils==7
+nutils==9
 numpy >1, <2
 pyprecice~=3.0
 matplotlib
