+++ partitioned-pipe-multiscale/fluid3d-left-openfoam/0/U
