+++ partitioned-pipe-multiscale/fluid3d-left-openfoam/clean.sh
