+++ partitioned-pipe-multiscale/fluid3d-right-openfoam/clean.sh
